🚚 divide endpoints to each api folder
