+++ src/api/queries/category/endpoint.js
@@ -1,0 +1,3 @@
+export const CATEGORY = `/categories`;
+export const FEED_LIST_BY_CATEGORY = (categoryId, cursorId) =>
+  `/categories/${categoryId || '1'}?cursorId=${cursorId || '0'}`;
