+++ src/api/queries/category/index.js
@@ -1,14 +1,14 @@
 import { api, apiWithToken } from 'api';
-import { CATEGORY, FEED_LIST_BY_CATEGORY } from 'api/endpoints';
+import * as EP from './endpoint';
 
 export const category = {
   getList: async () => {
-    const res = await api.get(CATEGORY);
+    const res = await api.get(EP.CATEGORY);
     return res.data;
   },
   getFeedListByCategory: async ({ categoryId, cursorId }) => {
     const res = await apiWithToken.get(
-      FEED_LIST_BY_CATEGORY(categoryId, cursorId)
+      EP.FEED_LIST_BY_CATEGORY(categoryId, cursorId)
     );
     return res.data;
   }
