+++ src/api/queries/comment/endpoint.js
@@ -1,0 +1,4 @@
+export const ADD_COMMENTS = feedId => `/comments/${feedId}`;
+export const ADD_REPLY = commentId => `/comments/reply/${commentId}`;
+export const GET_COMMENTS_LIST = (feedId, pageId) =>
+  `/comments/${feedId}?pageId=${pageId}`;
