+++ src/api/queries/feed/index.js
@@ -1,15 +1,9 @@
 import { apiWithToken } from 'api';
-import {
-  ADD_FEED,
-  GET_FEED_LIST,
-  GET_FEED_DETAIL,
-  BOOKMARK,
-  LIKE
-} from 'api/endpoints';
+import * as EP from './endpoint';
 
 export const feed = {
   add: async data => {
-    const res = await apiWithToken.post(ADD_FEED, data, {
+    const res = await apiWithToken.post(EP.ADD_FEED, data, {
       headers: {
         'Content-Type': 'multipart/form-data'
       }
@@ -18,29 +12,29 @@ export const feed = {
   },
   list: async ({ cursorId, pageId, sortType }) => {
     const res = await apiWithToken.get(
-      GET_FEED_LIST(cursorId, pageId, sortType)
+      EP.GET_FEED_LIST(cursorId, pageId, sortType)
     );
     return res.data;
   },
   detail: async ({ queryKey }) => {
     const { feedId } = queryKey[1];
-    const res = await apiWithToken.get(GET_FEED_DETAIL(feedId));
+    const res = await apiWithToken.get(EP.GET_FEED_DETAIL(feedId));
     return res.data;
   },
   addBookmark: async req => {
-    const res = await apiWithToken.post(BOOKMARK(req), req);
+    const res = await apiWithToken.post(EP.BOOKMARK(req), req);
     return res.data;
   },
   removeBookmark: async req => {
-    const res = await apiWithToken.delete(BOOKMARK(req));
+    const res = await apiWithToken.delete(EP.BOOKMARK(req));
     return res;
   },
   addLike: async req => {
-    const res = await apiWithToken.post(LIKE(req), req);
+    const res = await apiWithToken.post(EP.LIKE(req), req);
     return res.data;
   },
   removeLike: async req => {
-    const res = await apiWithToken.delete(LIKE(req));
+    const res = await apiWithToken.delete(EP.LIKE(req));
     return res;
   }
 };
