+++ src/api/queries/join/endpoint.js
@@ -1,0 +1,7 @@
+export const JOIN = `/users`;
+export const CHECK_NICKNAME = nickname =>
+  `/users/nickname?nickname=${nickname}`;
+
+export const SEND_CODE = `/users/verification-codes`;
+export const VERIFY_CODE = (authCode, phoneNumber) =>
+  `/users/verification-codes?authCode=${authCode}&phoneNumber=${phoneNumber}`;
